Split ui and routes; add stats proof of concept
